update locations translator
