+++ test/services/translators/geocode/LocationsTranslatorTest.js
@@ -1,0 +1,25 @@
+'use es6';
+
+import chai from 'chai';
+import sinon from 'sinon';
+import sinonChai from 'sinon-chai';
+chai.use(sinonChai);
+
+import LocationsTranslator from '../../../../src/services/translators/geocode/LocationsTranslator';
+
+const expect = chai.expect;
+
+describe('Locations Translation', function() {
+
+  describe('JSON Validation', () => {
+
+    describe('Valid', () => {
+
+    });
+
+    describe('Invalid', () => {
+  
+    });
+  });
+
+});
